% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_state_counts.R, R/cdc_states_counts.R
\docType{data}
\name{cdc_state_counts}
\alias{cdc_state_counts}
\title{Weekly death counts for each USA state}
\format{
An object of class \code{"data.frame"}

#' @docType data

An object of class \code{data.frame} with 9204 rows and 4 columns.
}
\usage{
data(cdc_state_counts)

data("cdc_state_counts")
}
\description{
The Center for Disease Control (CDC) provides weekly estimated death counts for each state in the USA. 
This data frame includes thes estimates for each state along with population sizes.
 
 \itemize{
 \item state Name of USA state
 \item date Corresponding date of observation
 \item outcome Estimated number of deaths
 \item population Population estimate (from the Census)}

A data frame with state-level weekley mortality counts from 2017 to 2020. The counts for 
the later weeks are adjusted for incomplete data using a method developed by CDC.
}
\keyword{datasets}
