% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_model.R
\name{excess_model}
\alias{excess_model}
\title{Fit excess count model}
\usage{
excess_model(
  counts,
  event = NULL,
  start = NULL,
  end = NULL,
  expected = NULL,
  exclude = NULL,
  trend.nknots = 1/5,
  harmonics = 2,
  family = "quasipoisson",
  day.effect = TRUE,
  nknots = 4,
  discontinuity = TRUE,
  max.iter = 15,
  eps = 1e-08,
  alpha = 0.05,
  min.rate = 0.01
)
}
\description{
Fit excess count model
}
