% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_stats.R
\name{excess_stats}
\alias{excess_stats}
\title{Compute excess deaths statistics for a list of intervals}
\usage{
excess_stats(
  counts,
  intervals,
  expected = NULL,
  exclude = NULL,
  trend.nknots = 1/5,
  harmonics = 2,
  family = "quasipoisson",
  day.effect = TRUE,
  max.control = 5000
)
}
\description{
Compute excess deaths statistics for a list of intervals
}
