% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ar.R
\name{fit_ar}
\alias{fit_ar}
\title{Fit an ar model to residuals from expected counts}
\usage{
fit_ar(counts, control.dates = NULL, order.max = 5, aic = FALSE, plot = FALSE)
}
\description{
Fit an ar model to residuals from expected counts
}
