% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_counts.R
\docType{data}
\name{florida_counts}
\alias{florida_counts}
\title{Florida daily mortality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1277 rows and 3 columns.
}
\usage{
data("florida_counts")
}
\description{
A data frame with Florida daily mortality counts from 2015-01-01 to 2018-06-30, which includes the day Irma made landfall on 2017-09-10.
The outcome column includes the number of deaths for that day.
}
\keyword{datasets}
