% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{Get demographic data from Census}
\usage{
get_demographics(
  geography = "state",
  state,
  county = NULL,
  years = 2018,
  vars = c("SEX", "AGEGROUP", "RACE", "HISP")
)
}
\description{
Get demographic data from Census
}
