% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_age.R
\name{group_age}
\alias{group_age}
\title{Assign age to group}
\usage{
group_age(age, breaks)
}
\arguments{
\item{age}{Vector of ages}

\item{breaks}{Ages that define strata}
}
\description{
Assign age to group
}
