% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisiana_counts.R
\docType{data}
\name{louisiana_counts}
\alias{louisiana_counts}
\title{Louisiana daily mortality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1461 rows and 3 columns.
}
\usage{
data("louisiana_counts")
}
\description{
A data frame with  Louisiana daily mortality counts from 2003-01-01 to 2016-12-31, 
which includes the day Katrina made landfall on 2015-08-29.
The outcome column includes the number of deaths for that day.
}
\keyword{datasets}
