% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_jersey_counts.R
\docType{data}
\name{new_jersey_counts}
\alias{new_jersey_counts}
\title{New Jersey daily mortality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3287 rows and 3 columns.
}
\usage{
data("new_jersey_counts")
}
\description{
A data frame with  Louisiana daily mortality counts from 2007-01-01 to 2015-12-31
which includes the day Sandy made landfall on 2012-10-29.
The outcome column includes the number of deaths for that day.
}
\keyword{datasets}
