% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puerto_rico_counts.R
\docType{data}
\name{puerto_rico_counts}
\alias{puerto_rico_counts}
\title{Puerto Rico daily mortality}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 499644 rows and 5 columns.
}
\usage{
data("puerto_rico_counts")
}
\description{
A data frame with Puerto Rico daily mortality counts, stratified by agegroup,
from 1985 to 2020.
which includes the days hurricanes Hugo, Georges, and Maria made
landfall on 1989-09-18, 1998-09-21, and 2017-09-20, respectively.
The outcome column includes the number of deaths for that day. This data frame
also includes population counts and estimates for the period. Population counts 
for 1985 to 2000 are interpolations of decennial census products since 1980 to 2000. 
For 2000 onward, we use interpolated population estimates (PEP).
}
\keyword{datasets}
