% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_counts.R
\docType{data}
\name{world_counts}
\alias{world_counts}
\title{Weekly death counts for several countries}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(world_counts)
}
\description{
A data frame with weekly death counts from multiple countries that includes the year 2020. 
Each country has a different range of data and most countries have at least 5 years of pre-2020 data.
The original data was collated by the Financial Times .
}
\details{
\itemize{
 \item country Name of the country
 \item date Corresponding date of observation
 \item outcome Estimated number of deaths
 \item population Population estimate (from the Census)}
}
\keyword{datasets}
