% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_demographics.R
\name{approx_demographics}
\alias{approx_demographics}
\title{Interpolate demographic data}
\usage{
approx_demographics(demo, first_day, last_day, by = c("year"))
}
\description{
Interpolate demographic data
}
