% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_demographics.R
\name{approx_demographics}
\alias{approx_demographics}
\title{Interpolate demographic data}
\usage{
approx_demographics(demo, first_day, last_day, by = "year")
}
\arguments{
\item{demo}{A data frame with the yearly population estimates}

\item{first_day}{First day to interpolate}

\item{last_day}{Last day to interpolate.}

\item{by}{Vector of column names to group by, for example different demographic strata}
}
\value{
A data frame with dates and population estimates.
}
\description{
Interpolate yearly population estimates so that a population estimate is provided for 
each day of the year. The function `approx` is used with `rule = 2` for extrapolation.
}
