% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_age_dist.R
\name{collapse_age_dist}
\alias{collapse_age_dist}
\alias{collapse_counts_by_age}
\title{Callapse age groups into broader ones}
\usage{
collapse_age_dist(demo, breaks)

collapse_counts_by_age(counts, breaks)
}
\arguments{
\item{demo}{A data frame with population sizes for different groups that will be collapsed}

\item{breaks}{The new age breaks for the new, broader, age strata}

\item{counts}{A data frame with counts and population sizes for different groups that will be collapsed}
}
\description{
Collapse a count or demographic data frame into a broader age strata.
}
\examples{
\dontrun{
data(cook_records)
demo <- collapse_age_dist(cook_demographics, 
                          breaks = c(0, 20, 40, 60, 80, Inf))
}
}
