% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_counts.R
\name{compute_counts}
\alias{compute_counts}
\title{Compute daily death counts for groups from individual records}
\usage{
compute_counts(
  dat,
  group_by = NULL,
  demo = NULL,
  date_column = "date",
  age_column = "age",
  agegroup_column = "agegroup",
  age_breaks = NULL
)
}
\description{
Compute daily death counts for groups from individual records
}
