% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expected.R
\name{compute_expected}
\alias{compute_expected}
\title{Compute expected counts for each day}
\usage{
compute_expected(
  counts,
  exclude = NULL,
  trend.nknots = 1/5,
  harmonics = 2,
  family = "quasipoisson",
  day.effect = TRUE
)
}
\description{
Compute expected counts for each day
}
