% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_plot.R
\name{excess_plot}
\alias{excess_plot}
\title{Plot results from fitted excess count model}
\usage{
excess_plot(fit, title = "", ylim = NULL, show.data = TRUE, alpha = 0.05)
}
\description{
Plot results from fitted excess count model
}
