% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_stats.R
\name{excess_stats}
\alias{excess_stats}
\title{Compute excess deaths statistics for a list of intervals}
\usage{
excess_stats(
  counts,
  intervals,
  expected = NULL,
  exclude = NULL,
  model = c("quasipoisson", "poisson", "correlated"),
  trend.nknots = 1/5,
  harmonics = 2,
  day.effect = TRUE,
  control.dates = NULL,
  max.control = 5000,
  order.max = 30,
  aic = TRUE,
  verbose = TRUE
)
}
\description{
Compute excess deaths statistics for a list of intervals
}
