% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_stats.R
\name{excess_stats}
\alias{excess_stats}
\title{Helper funtion to compute excess deaths statistics for a}
\usage{
excess_stats(
  start,
  end,
  obs,
  mu,
  cov,
  pop,
  frequency,
  fhat = NULL,
  X = NULL,
  betacov = NULL
)
}
\description{
Helper funtion to compute excess deaths statistics for a
}
