% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_plot.R
\name{expected_plot}
\alias{expected_plot}
\title{Check if expected counts fit data}
\usage{
expected_plot(
  expected,
  title = "",
  start = NULL,
  end = NULL,
  ylim = NULL,
  weekly = TRUE,
  color = "#3366FF",
  alpha = 1
)
}
\description{
Check if expected counts fit data
}
