% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demographics.R
\name{get_demographics}
\alias{get_demographics}
\title{Get demographic data from Census}
\usage{
get_demographics(
  geography = "state",
  state,
  county = NULL,
  years = 2018,
  vars = c("SEX", "AGEGROUP", "RACE", "HISP")
)
}
\arguments{
\item{geography}{state or county}

\item{state}{name of the state}

\item{county}{name of the county}

\item{years}{vector of years for which we obtain data}

\item{vars}{names of variables that define strat of which we want population estimates}
}
\description{
Get demographic data from Census
}
